ToolCalling class
